+++ src/core/LiveWorkItemDetails.ts
@@ -3,5 +3,7 @@ import LiveWorkItemState from "./LiveWorkItemState";
 export default interface LiveWorkItemDetails {
     state: LiveWorkItemState,
     agentAcceptedOn?: string,
-    conversationId: string
+    conversationId: string,
+    canRenderPostChat?: string,
+    participantType?: string
 }