+++ src/core/PostChatContext.ts
@@ -1,5 +1,6 @@
 export default interface PostChatContext {
   formsProLocale: string;
   participantJoined: boolean;
+  participantType: string | undefined;
   surveyInviteLink: string;
 }