+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 - Add `acs_webchat-chat-adapter` middleware to add default `channelData.tags` & `channelData.metadata`
 - Update `ChatConfig` interface with `msdyn_enablemarkdown` property
 - Throw exception on `ChatSDK.getVoiceVideoCalling()` if feature is disabled or platform is not supported
+- Add `participantType` & `canRenderPostChat` as response of getConversationDetails() API
+- Add support for separate bot post chat survey feature
 
 ### Fix
 - Add `acs_webchat-chat-adapter` middlewares to format `channelData.tags`